backend/pyproject.toml
[project]
name = "fundsys-project"
version = "0.2.0"
description = "Projeto criado para realizar o teste técnico da Fundsys"
authors = [
    {name = "lucaspbueno",email = "lucaspbueno22@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.16.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "requests (>=2.31.0,<3.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
